Add new persona for the manufacturing area
Withe the manufacturing area we have the opportunity to explore the use of IoT sensors in the manufacturing plant.  Also the processes and operation of the manufacturing plant will be effected by the move to personalized medicine so Stew will play a role in that - particularly as a stakeholder in the IT / data provision for the manufacturing part of the business.